watch: Take failing readers off the readers list
As a temporary hack we take failing read callbacks off the readers list.

When this happens proper cleanup of the read watcher should be done, so
this needs to be revised further. This is necessary when implementing a
UNIX socket server, as the read operation will fail when the clients
disconnect.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>