perif-rpmsg: Only initialize on primary channel
Rather than initiating the opening of a peripheral on any channel on a
new remoteproc we look only for the primary channels (DIAG or
APPS_RIVA_DATA). This solves a problem seen on MSM8996 where the DIAG
channels somtimes are created with a significant delay from the first
channels appearing, causing the 1 second delay to not be enough.

We keep the 1 second timer before trying to open the peripheral as this
forms a wait for the seconday channels to be registered. An alternative
would be to retry until the channels are present, but as the CMD channel
is optional there isn't a good way to detect if this is absent of not
yet registered - potentially this could be solved by opening it lazily
if the remote reports REQ_RSP support.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>