usb: Fix USB implementation
USB support is implemented using FunctionFS, which requires that one
uses AIO in order to support non-blocking operation. The previously
introduced change in watch.c takes care of the IN ep, but OUT is now
moved to use the new AIO read support.

Also remove all traces of Gadget ConfigFS configuration, as this belongs
in the system configuration, rather than in the diag tool itself.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>