watch: Allow flow context to be associated with readq
Allow a flow context to be associated to a read queue and consider the
flow blockage when deciding to evaluate the associated file descriptor
for incoming data.

This allows clients to associate a read queue with the same flow context
as they will associate mbufs to be flow controlled with.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>