peripherals: Create and use flow context
Introduce a flow control context per peripheral and pass this to
dm_broadcast() in order for any queued messages to be accounted for in
the flow control context. Associate the same flow control context with
the read watcher for the peripheral's incoming data descriptors to make
the watch runner selectively read from the given peripheral only when
the number of outstanding, queued, messages are below the threshold.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>